+++ compose-ui/src/commonMain/kotlin/com/tlozovyi/weekview/WeekViewPagingEffect.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2026 Taras Lozovyi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.tlozovyi.weekview
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import kotlinx.datetime.LocalDate
+
+@Composable
+internal fun WeekViewPagingEffect(
+    pagingState: WeekViewPagingState?,
+    settlementGeneration: Int,
+    firstVisibleDate: LocalDate,
+    numberOfVisibleDays: Int,
+    isLtr: Boolean,
+) {
+    if (pagingState == null) {
+        return
+    }
+
+    LaunchedEffect(pagingState, settlementGeneration, firstVisibleDate, numberOfVisibleDays, isLtr) {
+        val (rangeStart, rangeEnd) = visibleDateRange(
+            firstVisibleDate = firstVisibleDate,
+            numberOfVisibleDays = numberOfVisibleDays,
+            isLtr = isLtr,
+        )
+        pagingState.onScrollSettled(
+            firstVisibleDate = rangeStart,
+            lastVisibleDate = rangeEnd,
+        )
+    }
+}